Minor: move test cases from TimeConversionTest to ConsumerGroupCommandServerlessTest
